Develop (#32)
* Modify develop dependency


* Fix error in FallingSand project


* Fix error in func getRGB565BE


---------

+++ Examples/SwiftIOPlayground/08LCD/Rainbow/Sources/main.swift
@@ -47,5 +47,5 @@ while true {
 // The screen needs RGB565 color data, so change color data from UInt32 to UInt16.
 // Besides, the board uses little endian format, so the bytes are swapped.
 func getRGB565BE(_ color: UInt32) -> UInt16 {
-    return UInt16(((color & 0xF80000) >> 8) | ((color & 0xFC00) >> 5) | ((color & 0xF8) >> 3)).byteSwapped
+    return (UInt16((color & 0xF80000) >> 8) | UInt16((color & 0xFC00) >> 5) | UInt16((color & 0xF8) >> 3)).byteSwapped
 }
